Update image-flow extension
- remove unused code
- fix PR comments that when submit to Raycast store ^_-
- add validate when output is iamge or not
- move process login to one file
- New process all command
- wip
- update demo video

+++ extensions/image-flow/src/actions/rename.ts
@@ -26,7 +26,9 @@ export default async function (i: Input, config: Config): Promise<Output> {
   const newPath = path.join(path.dirname(inputPath), newName);
 
   await saveStreamToFile(fs.createReadStream(inputPath), newPath);
-  fs.unlinkSync(inputPath);
+  if (fs.existsSync(inputPath)) {
+    fs.unlinkSync(inputPath);
+  }
 
   return { type: "filepath", value: newPath } as Output;
 }
